requirements.txt
Flask==2.1.1
gunicorn==20.1.0
python-dotenv==0.20.0